Refactors Search Bar.
